fix GameUpdate
